modify trainer folder to work with citibike dataset
